rename modal.tsx in backend to dbCommands. debug undefined or null values in DataTable
